Fitur 7 bisa, fitur 8 bisa request tapi belom bisa masuk (revisi 2), fitur 9 done

+++ TA_6/src/main/java/com/apap/TA/repository/JadwalJagaDb.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.apap.TA.model.JadwalJagaModel;
 
 public interface JadwalJagaDb extends JpaRepository <JadwalJagaModel,  Long>{
-	Optional <JadwalJagaModel> findById (long id);
+	Optional <JadwalJagaModel> findById (int id);
 	Optional <JadwalJagaModel> findByTanggal (Date tanggal);
 
 }
